requirements.txt
asn1crypto==1.5.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
distlib==0.3.6
filelock==3.12.1
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
platformdirs==3.5.3
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
Werkzeug==2.3.6
